+++ packages/f-services/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 1.0.0
 ------------------------------
-*March 30, 2020*
+*April 1, 2020*
 
 ### Added
 - Axios services module.
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Split central "shared" module into separate "window" and "globalisation" modules.
 - `jest` config updates. `moduleNameMapper` removed, `jsx` removed from checks (as we don't use) and `transformIgnorePatterns` updated.
 - Rollup package updates.
+
+### Fixed
+- `addEvent` returns the function that was assigned to the event listener so that this can be passed in to `removeEvent`. This is necessary for removing throttled event listeners.
 
 
 0.13.2